add example for short deck and synced calculations
